Separate DaCeStencilObject form StencilObject
