+++ src/gt4py/stencil_object.py
@@ -16,35 +16,24 @@
 
 import abc
 import collections.abc
-import copy
-import inspect
-import os
 import sys
 import time
 import typing
 import warnings
 from dataclasses import dataclass
 from pickle import dumps
-from typing import Any, Callable, ClassVar, Dict, Optional, Sequence, Set, Tuple, Union
+from typing import Any, Callable, ClassVar, Dict, Optional, Tuple, Union
 
-import dace
-import dace.data
-import dace.frontend.python.common
 import numpy as np
-from dace.frontend.python.common import SDFGClosure, SDFGConvertible
-from dace.sdfg.utils import inline_sdfgs
 
 import gt4py.backend as gt_backend
 import gt4py.storage as gt_storage
 import gt4py.utils as gt_utils
-from gt4py.definitions import DomainInfo, FieldInfo, Index, ParameterInfo, Shape
-from gt4py.utils import shash
+from gt4py.definitions import AccessKind, DomainInfo, FieldInfo, Index, ParameterInfo, Shape
 
 
 FieldType = Union[gt_storage.storage.Storage, np.ndarray]
 OriginType = Union[Tuple[int, int, int], Dict[str, Tuple[int, ...]]]
-
-_loaded_sdfgs: Dict[str, dace.SDFG] = dict()
 
 
 def _compute_cache_key(field_args, parameter_args, domain, origin) -> int:
@@ -58,7 +47,7 @@ def _compute_cache_key(field_args, parameter_args, domain, origin) -> int:
 
 
 @dataclass(frozen=True)
-class FrozenStencil(SDFGConvertible):
+class FrozenStencil:
     """Stencil with pre-computed domain and origin for each field argument."""
 
     stencil_object: "StencilObject"
@@ -67,9 +56,7 @@ class FrozenStencil(SDFGConvertible):
 
     def __post_init__(self):
         for name, field_info in self.stencil_object.field_info.items():
-            if field_info is not None and (
-                name not in self.origin or len(self.origin[name]) != field_info.ndim
-            ):
+            if name not in self.origin or len(self.origin[name]) != field_info.ndim:
                 raise ValueError(
                     f"'{name}' origin {self.origin.get(name)} is not a {field_info.ndim}-dimensional integer tuple"
                 )
@@ -95,210 +82,8 @@ class FrozenStencil(SDFGConvertible):
         if exec_info is not None:
             exec_info["call_run_end_time"] = time.perf_counter()
 
-    def _sdfg_add_arrays_and_edges(self, wrapper_sdfg, state, inner_sdfg, nsdfg, inputs, outputs):
-        device = gt_backend.from_name(self.stencil_object.backend).storage_info["device"]
 
-        for name, array in inner_sdfg.arrays.items():
-            if isinstance(array, dace.data.Array) and not array.transient:
-                axes = self.stencil_object.field_info[name].axes
-
-                shape = [f"__{name}_{axis}_size" for axis in axes] + [
-                    str(d) for d in self.stencil_object.field_info[name].data_dims
-                ]
-
-                wrapper_sdfg.add_array(
-                    name,
-                    dtype=array.dtype,
-                    strides=array.strides,
-                    shape=shape,
-                    storage=dace.StorageType.GPU_Global
-                    if device == "gpu"
-                    else dace.StorageType.Default,
-                )
-                if isinstance(self.origin, tuple):
-                    origin = [o for a, o in zip("IJK", self.origin) if a in axes]
-                else:
-                    origin = self.origin.get(name, self.origin.get("_all_", None))
-                    if len(origin) == 3:
-                        origin = [o for a, o in zip("IJK", origin) if a in axes]
-
-                subset_strs = [
-                    f"{o - e}:{o - e + s}"
-                    for o, e, s in zip(
-                        origin,
-                        self.stencil_object.field_info[name].boundary.lower_indices,
-                        inner_sdfg.arrays[name].shape,
-                    )
-                ]
-                subset_strs += [f"0:{d}" for d in self.stencil_object.field_info[name].data_dims]
-                if name in inputs:
-                    state.add_edge(
-                        state.add_read(name),
-                        None,
-                        nsdfg,
-                        name,
-                        dace.Memlet.simple(name, ",".join(subset_strs)),
-                    )
-                if name in outputs:
-                    state.add_edge(
-                        nsdfg,
-                        name,
-                        state.add_write(name),
-                        None,
-                        dace.Memlet.simple(name, ",".join(subset_strs)),
-                    )
-
-    def _sdfg_specialize_symbols(self, wrapper_sdfg):
-        ival, jval, kval = self.domain[0], self.domain[1], self.domain[2]
-        for sdfg in wrapper_sdfg.all_sdfgs_recursive():
-            if sdfg.parent_nsdfg_node is not None:
-                symmap = sdfg.parent_nsdfg_node.symbol_mapping
-
-                if "__I" in symmap:
-                    ival = symmap["__I"]
-                    del symmap["__I"]
-                if "__J" in symmap:
-                    jval = symmap["__J"]
-                    del symmap["__J"]
-                if "__K" in symmap:
-                    kval = symmap["__K"]
-                    del symmap["__K"]
-
-            sdfg.replace("__I", ival)
-            if "__I" in sdfg.symbols:
-                sdfg.remove_symbol("__I")
-            sdfg.replace("__J", jval)
-            if "__J" in sdfg.symbols:
-                sdfg.remove_symbol("__J")
-            sdfg.replace("__K", kval)
-            if "__K" in sdfg.symbols:
-                sdfg.remove_symbol("__K")
-
-            for val in ival, jval, kval:
-                sym = dace.symbolic.pystr_to_symbolic(val)
-                for fsym in sym.free_symbols:
-                    if sdfg.parent_nsdfg_node is not None:
-                        sdfg.parent_nsdfg_node.symbol_mapping[str(fsym)] = fsym
-                    if str(fsym) not in sdfg.symbols:
-                        if str(fsym) in sdfg.parent_sdfg.symbols:
-                            sdfg.add_symbol(str(fsym), stype=sdfg.parent_sdfg.symbols[str(fsym)])
-                        else:
-                            sdfg.add_symbol(str(fsym), stype=dace.dtypes.int32)
-
-    def _sdfg_freeze_domain_and_origin(self, inner_sdfg: dace.SDFG):
-        wrapper_sdfg = dace.SDFG("frozen_" + inner_sdfg.name)
-        state = wrapper_sdfg.add_state("frozen_" + inner_sdfg.name + "_state")
-
-        inputs = set()
-        outputs = set()
-        for inner_state in inner_sdfg.nodes():
-            for node in inner_state.nodes():
-                if (
-                    not isinstance(node, dace.nodes.AccessNode)
-                    or inner_sdfg.arrays[node.data].transient
-                ):
-                    continue
-                if node.has_reads(inner_state):
-                    inputs.add(node.data)
-                if node.has_writes(inner_state):
-                    outputs.add(node.data)
-
-        nsdfg = state.add_nested_sdfg(inner_sdfg, None, inputs, outputs)
-
-        self._sdfg_add_arrays_and_edges(wrapper_sdfg, state, inner_sdfg, nsdfg, inputs, outputs)
-
-        # in special case of empty domain, remove entire SDFG.
-        if any(d == 0 for d in self.domain):
-            states = wrapper_sdfg.states()
-            assert len(states) == 1
-            for node in states[0].nodes():
-                state.remove_node(node)
-
-        # make sure that symbols are passed throught o inner sdfg
-        for symbol in nsdfg.sdfg.free_symbols:
-            if symbol not in wrapper_sdfg.symbols:
-                wrapper_sdfg.add_symbol(symbol, nsdfg.sdfg.symbols[symbol])
-
-        # Try to inline wrapped SDFG before symbols are specialized to avoid extra views
-        inline_sdfgs(wrapper_sdfg)
-
-        self._sdfg_specialize_symbols(wrapper_sdfg)
-
-        for _, _, array in wrapper_sdfg.arrays_recursive():
-            if array.transient:
-                array.lifetime = dace.dtypes.AllocationLifetime.SDFG
-
-        signature = self.__sdfg_signature__()
-        wrapper_sdfg.arg_names = [a for a in signature[0] if a not in signature[1]]
-
-        return wrapper_sdfg
-
-    def _assert_dace_backend(self):
-        if not hasattr(self.stencil_object, "_sdfg"):
-            raise TypeError(
-                f"Only dace backends are supported in DaCe-orchestrated programs."
-                f' (found "{self.stencil_object.backend}")'
-            )
-
-    def _add_optionals(self, sdfg, **kwargs):
-        for name, info in self.stencil_object.field_info.items():
-            if name in kwargs and (info is None or name not in sdfg.arrays):
-                outer_array = kwargs[name]
-                sdfg.add_array(
-                    name,
-                    shape=outer_array.shape,
-                    dtype=outer_array.dtype,
-                    strides=outer_array.strides,
-                )
-
-        for name, info in self.stencil_object.parameter_info.items():
-            if name in kwargs and (info is None or name not in sdfg.symbols):
-                if isinstance(kwargs[name], dace.data.Scalar):
-                    sdfg.add_symbol(name, stype=kwargs[name].dtype)
-                else:
-                    sdfg.add_symbol(name, stype=dace.typeclass(type(kwargs[name])))
-        return sdfg
-
-    def __sdfg__(self, **kwargs):
-
-        self._assert_dace_backend()
-        frozen_hash = shash(type(self.stencil_object)._gt_id_, self.origin, self.domain)
-
-        # check if same sdfg already cached in memory
-        if frozen_hash in _loaded_sdfgs:
-            return self._add_optionals(copy.deepcopy(_loaded_sdfgs[frozen_hash]), **kwargs)
-
-        # check if same sdfg already cached on disk
-        basename = os.path.splitext(self.stencil_object._file_name)[0]
-        filename = basename + "_" + str(frozen_hash) + ".sdfg"
-        try:
-            _loaded_sdfgs[frozen_hash] = dace.SDFG.from_file(filename)
-            return self._add_optionals(copy.deepcopy(_loaded_sdfgs[frozen_hash]), **kwargs)
-        except FileNotFoundError:
-            pass
-
-        # otherwise, wrap and save sdfg from scratch
-        inner_sdfg = self.stencil_object.sdfg
-
-        _loaded_sdfgs[frozen_hash] = self._sdfg_freeze_domain_and_origin(inner_sdfg)
-        _loaded_sdfgs[frozen_hash].save(filename)
-
-        return self._add_optionals(copy.deepcopy(_loaded_sdfgs[frozen_hash]), **kwargs)
-
-    def __sdfg_signature__(self):
-        self._assert_dace_backend()
-        return self.stencil_object.__sdfg_signature__()
-
-    def __sdfg_closure__(self, *args, **kwargs):
-        self._assert_dace_backend()
-        return {}
-
-    def closure_resolver(self, constant_args, given_args, parent_closure=None):
-        self._assert_dace_backend()
-        return SDFGClosure()
-
-
-class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
+class StencilObject(abc.ABC):
     """Generic singleton implementation of a stencil callable.
 
     This class is used as base class for specific subclass generated
@@ -343,7 +128,6 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
     # Those attributes are added to the class at loading time:
     _gt_id_: str
     definition_func: Callable[..., Any]
-    _frozen_cache: Dict[Any, FrozenStencil]
 
     _domain_origin_cache: ClassVar[Dict[int, Tuple[Tuple[int, ...], Dict[str, Tuple[int, ...]]]]]
     """Stores domain/origin pairs that have been used by hash."""
@@ -351,7 +135,7 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
     def __new__(cls, *args, **kwargs):
         if getattr(cls, "_instance", None) is None:
             cls._instance = object.__new__(cls)
-            cls._frozen_cache = dict()
+            cls._domain_origin_cache = {}
         return cls._instance
 
     def __setattr__(self, key, value) -> None:
@@ -474,7 +258,7 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
         max_domain = Shape([max_size] * domain_ndim)
 
         for name, field_info in self.field_info.items():
-            if field_info is not None:
+            if field_info.access != AccessKind.NONE:
                 assert field_args.get(name, None) is not None, f"Invalid value for '{name}' field."
                 field = field_args[name]
                 api_domain_mask = field_info.domain_mask
@@ -530,9 +314,14 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
                 f"Compute domain too large (provided: {domain}, maximum: {max_domain})"
             )
 
+        if domain[2] < self.domain_info.min_sequential_axis_size:
+            raise ValueError(
+                f"Compute domain too small. Sequential axis is {domain[2]}, but must be at least {self.domain_info.min_sequential_axis_size}."
+            )
+
         # assert compatibility of fields with stencil
         for name, field_info in self.field_info.items():
-            if field_info is not None:
+            if field_info.access != AccessKind.NONE:
                 if name not in field_args:
                     raise ValueError(f"Missing value for '{name}' field.")
                 field = field_args[name]
@@ -603,9 +392,10 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
                     )
 
                 spatial_domain = typing.cast(Shape, domain).filter_mask(field_domain_mask)
+                lower_indices = field_info.boundary.lower_indices.filter_mask(field_domain_mask)
                 upper_indices = field_info.boundary.upper_indices.filter_mask(field_domain_mask)
                 min_shape = tuple(
-                    o + d + h for o, d, h in zip(field_domain_origin, spatial_domain, upper_indices)
+                    lb + d + ub for lb, d, ub in zip(lower_indices, spatial_domain, upper_indices)
                 )
                 if min_shape > field.shape:
                     raise ValueError(
@@ -614,10 +404,10 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
 
         # assert compatibility of parameters with stencil
         for name, parameter_info in self.parameter_info.items():
-            if parameter_info is not None:
+            if parameter_info.access != AccessKind.NONE:
                 if name not in param_args:
                     raise ValueError(f"Missing value for '{name}' parameter.")
-                if type(parameter := param_args[name]) != parameter_info.dtype:
+                elif type(parameter := param_args[name]) != parameter_info.dtype:
                     raise TypeError(
                         f"The type of parameter '{name}' is '{type(parameter)}' instead of '{parameter_info.dtype}'"
                     )
@@ -630,7 +420,7 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
 
         # Set an appropriate origin for all fields
         for name, field_info in self.field_info.items():
-            if field_info is not None:
+            if field_info.access != AccessKind.NONE:
                 assert name in field_args, f"Missing value for '{name}' field."
                 field_origin = origin.get(name, None)
 
@@ -648,8 +438,7 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
                         *((0,) * len(field_info.data_dims)),
                     )
 
-                elif hasattr(field_arg := field_args.get(name), "default_origin"):
-                    assert field_arg is not None
+                elif isinstance(field_arg := field_args.get(name), gt_storage.storage.Storage):
                     origin[name] = field_arg.default_origin
 
                 else:
@@ -699,13 +488,19 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
         if exec_info is not None:
             exec_info["call_run_start_time"] = time.perf_counter()
 
-        origin = self._normalize_origins(field_args, origin)
+        cache_key = _compute_cache_key(field_args, parameter_args, domain, origin)
+        if cache_key not in self._domain_origin_cache:
+            origin = self._normalize_origins(field_args, origin)
 
-        if domain is None:
-            domain = self._get_max_domain(field_args, origin)
+            if domain is None:
+                domain = self._get_max_domain(field_args, origin)
 
-        if validate_args:
-            self._validate_args(field_args, parameter_args, domain, origin)
+            if validate_args:
+                self._validate_args(field_args, parameter_args, domain, origin)
+
+            type(self)._domain_origin_cache[cache_key] = (domain, origin)
+        else:
+            domain, origin = type(self)._domain_origin_cache[cache_key]
 
         self.run(
             _domain_=domain, _origin_=origin, exec_info=exec_info, **field_args, **parameter_args
@@ -713,11 +508,6 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
 
         if exec_info is not None:
             exec_info["call_run_end_time"] = time.perf_counter()
-
-    @staticmethod
-    def _get_domain_origin_key(domain, origin):
-        origins_tuple = tuple((k, v) for k, v in sorted(origin.items()))
-        return domain, origins_tuple
 
     def freeze(
         self: "StencilObject", *, origin: Dict[str, Tuple[int, ...]], domain: Tuple[int, ...]
@@ -745,54 +535,7 @@ class StencilObject(abc.ABC, dace.frontend.python.common.SDFGConvertible):
                 or domain occurs at call time so it is the users responsibility to ensure
                 correct usage.
         """
-        key = StencilObject._get_domain_origin_key(domain, origin)
-        if key not in self._frozen_cache:
-            self._frozen_cache[key] = FrozenStencil(self, origin, domain)
-        return self._frozen_cache[key]
-
-    def _normalize_args(self, *args, domain=None, origin=None, **kwargs):
-        arg_names, consts = self.__sdfg_signature__()
-        args_iter = iter(args)
-        args_as_kwargs = {
-            name: (kwargs[name] if name in kwargs else next(args_iter)) for name in arg_names
-        }
-
-        origin = self._normalize_origins(args_as_kwargs, origin)
-        if domain is None:
-            domain = self._get_max_domain(args_as_kwargs, origin)
-        for key, value in kwargs.items():
-            args_as_kwargs.setdefault(key, value)
-        args_as_kwargs["domain"] = domain
-        args_as_kwargs["origin"] = origin
-        return args_as_kwargs
-
-    def __sdfg__(self, *args, **kwargs) -> dace.SDFG:
-        norm_kwargs = self._normalize_args(*args, **kwargs)
-        frozen_stencil = self.freeze(origin=norm_kwargs["origin"], domain=norm_kwargs["domain"])
-        return frozen_stencil.__sdfg__(**norm_kwargs)
-
-    def __sdfg_closure__(self, reevaluate: Optional[Dict[str, str]] = None) -> Dict[str, Any]:
-        return {}
-
-    def __sdfg_signature__(self) -> Tuple[Sequence[str], Sequence[str]]:
-        special_args = {"self", "domain", "origin", "validate_args", "exec_info"}
-        args = []
-        for arg in (
-            inspect.getfullargspec(self.__call__).args
-            + inspect.getfullargspec(self.__call__).kwonlyargs
-        ):
-            if arg in special_args:
-                continue
-            args.append(arg)
-        return (args, [])
-
-    def closure_resolver(
-        self,
-        constant_args: Dict[str, Any],
-        given_args: Set[str],
-        parent_closure: Optional["dace.frontend.python.common.SDFGClosure"] = None,
-    ) -> "dace.frontend.python.common.SDFGClosure":
-        return dace.frontend.python.common.SDFGClosure()
+        return FrozenStencil(self, origin, domain)
 
     def clean_call_args_cache(self: "StencilObject") -> None:
         """Clean the argument cache.
